fix(types): address dummy app review cleanup
