+++ react_on_rails/spec/dummy/client/app/startup/ReduxSharedStoreApp.client.tsx
@@ -1,35 +1,11 @@
-import React, { type ReactNode } from 'react';
-import ReactDOMClient from 'react-dom/client';
+import React from 'react';
 import { Provider } from 'react-redux';
 import ReactOnRails from 'react-on-rails/client';
 
 import HelloWorldContainer from '../components/HelloWorldContainer';
 import { wrapElementInStrictMode } from '../strictModeSupport';
 import type { ReduxAppStore } from '../store/reduxTypes';
-
-type DomRenderer = (domNode: Element, element: ReactNode) => void;
-type ReactRoot = ReturnType<typeof ReactDOMClient.createRoot>;
-
-const reactRoots = new WeakMap<Element, ReactRoot>();
-
-const hydrateOrRender =
-  (shouldHydrate: boolean): DomRenderer =>
-  (domNode, element) => {
-    const existingRoot = reactRoots.get(domNode);
-    if (existingRoot) {
-      existingRoot.render(element);
-      return;
-    }
-
-    const root = shouldHydrate
-      ? ReactDOMClient.hydrateRoot(domNode, element)
-      : ReactDOMClient.createRoot(domNode);
-    reactRoots.set(domNode, root);
-
-    if (!shouldHydrate) {
-      root.render(element);
-    }
-  };
+import { hydrateOrRender } from './domRenderers';
 
 export default function ReduxSharedStoreApp(
   props: Record<string, unknown>,
